Log company that is currently being uploaded in the deploy script to avoid Travis killing the build
We have reached a number of records that causes the upload to take longer than 10 minutes.
Travis automatically kills jobs that run longer than 10 minutes without producing any output
(see: https://docs.travis-ci.com/user/common-build-problems/#build-times-out-because-no-output-was-received).

Instead of relying on the Travis provided `travis_wait` tool to produce dummy output in the
background, we might as well log the record (slug) we are currently working on.