Most companies won't need the birthdate for requests
The date of birth is only necessary is only necessary if a request
is otherwise ambiguous (e.g. if two people with the same name live
at the same address and both have an account with the company in
question). This should only happen very rarely.
Thus, there is no reason to always provide that data. Instead, the
companies will have to ask for it where necessary.