Make German greeting more respectful and inclusive
One of the replies I've got criticized that the greeting "Guten Tag"
is disrespectful. A greeting like "Sehr geehrte Damen und Herren" may
be more appropiate. To also include people that neither identify as
"Damen" nor "Herren", I've added the asterisk (*) symbol. It is
commonly used to refer to other kinds of people.