Fix email of IFTTT Inc.
The email in the terms is "IFTTT@gdpr-rep.com".
However, when sending a request to that email, you get the
following reply:
"We don't monitor emails sent directly to this email address, and
we won't be able to respond to this email."
Therefore, I've added the more general email address that can also
be found in the sources.