Migrate the record test script from Python to JS
The primary reason here is that our new Xapiand deploy script will
be written in JS, too and it is better to be consistent.

The new JSON schema validation library we are using is also more
comprehensive (it also checks the `format`).

Finally, the new version is also about 10 times quicker, which is
nice.