Created template to access according to the law on church data protection for the catholic church
According to Art. 91 GDPR, churches and other religious associations have the right to issue own privacy regulations for their internal affairs, provided that they are in line with GDPR regulations. Thus the rights of the data subjects need to be exercised in accordance with   the privacy law of the respective church or religious association. This file contains the right to access request pursuant the law on church data protection (Gesetz über den Kirchlichen Datenschutz) for the catholic church in Germany.