Add templates for requests to the catholic church in Germany (#451, #565)
According to Art. 91 GDPR, churches and other religious associations have the right to issue own privacy regulations for their internal affairs, provided that they are in line with GDPR regulations. Thus the rights of the data subjects need to be exercised in accordance with the privacy law of the respective church or religious association.

This commit adds the following templates:

- access
- admonition
- complaint
- erasure
- objection
- rectification

Co-authored-by: Benjamin Altpeter <benjamin@oldpeter.de>