Create access-evangelische-kirche.txt
The DSG-EKD does not contain a right to obtain a copy of the processed date (Art. 15(3) GDPR). Furthermore there is no right to obtain information about automated decision-making (Art. 15(1) point (h) GDPR), and no right to obtain information about suitable safeguards, when personal data are transferred to a third country (Art. 15(2) GDPR). Finally the deadline to answer the request is set to three month instead of one.
As these provisions are likely not in line with GDPR (Art. 91(2) GDPR), I personally would at least request a copy of the processed information anyway. The legality of such provisions is at least questionable.
The questionable sections are still in the template, since I am a bit unsure what to do with them.