Add Tinder
It's unfortunate that they do not even list a mail in their privacy policy: https://policies.tinder.com/privacy/

Also the Match Group has a lot of more companies that belong to them, but they do partially have separate privacy policies and contacts so let's not include OkCupid etc. in here.

As for `suggested-transport-medium` I would suggest to use the web form, but that is not really possible.