Add templates for the protestant church in Germany. (#571)
The DSG-EKD does not contain a right to obtain a copy of the processed date (Art. 15(3) GDPR). Furthermore there is no right to obtain information about automated decision-making (Art. 15(1) point (h) GDPR), and no right to obtain information about suitable safeguards, when personal data are transferred to a third country (Art. 15(2) GDPR). Finally the deadline to answer the request is set to three month instead of one.

Co-authored-by: Benjamin Altpeter <benjamin@oldpeter.de>