Add all churches from the Landeskirche Braunschweig
This data was automatically scraped using the script in the
datenanfragen/data-imports repository.

The data was scraped from the subpages of this website:
https://www.landeskirche-braunschweig.de/gemeinden/propsteien/

A handful of email addresses containing personal data were manually
swapped out.
In addition, the email addresses for the "Propsteien" were added
manually as they are 'protected' on the site and it didn't make
sense to automate this process considering there are only twelve
of them.