Update outdated email address (letters A and B)
As it turns out, our data is rotting a lot faster than I
expected.

This is the first batch of manual updates based on a tool
first proposed and implemented by @mal-tee that I have now
taken over.
So far, I have only completed the results for companies with
slugs starting with A and B, and the tool currently only checks
the email addresses.

Clearly, this is an important problem that we will need to deal
with. The ultimate goal for the tool is to automatically run it
periodically and report its findings as issues on this repo. This
is however not implemented yet but I will prioritize working on it.
I will also manually handle the first batch of findings mysel as
we would otherwise be completely overwhelmed with issues.
Hopefully, this will become more manageable once we actually
start regularly checking everything.