Clarify that we are requesting a copy in access templates (DE + EN) (#1167)
In the subject literature, there is a bit of a divide concerning the
relationship between Art. 15(1) and (3) GDPR. Basically, those are the
two differing opinions:

1. The "The controller shall provide a copy of the personal data
   undergoing processing." from Art. 15(3) GDPR mandates how, in
   particular, to give the data subject "access to the personal data"
   according to Art. 15(1) GDPR.

   As such, a request according to Art. 15(1) GDPR includes a request
   for a copy.
2. Art. 15(1) and (3) GDPR are separate rights. A data subject that
   wants a copy of their data explicitly has to request that according
   to Art. 15(3) GDPR.

We believe that by requesting "all personal data concerning me that
you have stored", as we did until now in our access template, it is
obvious that we are in fact requesting a copy of said data.
From countless access requests using our template that we have
submitted ourselves or have been told about by our users, we have also
not seen a single case where a controller interpreted our wording
differently.

However, a user recently kindly forwarded us a response from a data
protection authority to a complaint that did not provide a copy after
they requested access using our template. The authority ruled that it
was not clear from the request that the user in fact asked for a copy
and as such the controller acted correctly.

While we don't agree with the authority's assessment, one of our
primary goals with our request templates is to provide legal certainty
to our users.
As such, we have decided to change the access templates to explicitly
ask for a copy according to Art. 15(3) GDPR. This ensures that there
can be no more misunderstandings in the future and that our users have
a strong legal case for defending against controllers that don't
provide them with a copy of their data.

For reference, this is some of the literature on this topic (only in
German, unfortunately):

* BeckOK DatenschutzR/Schmidt-Wudy, 34. Ed. 1.11.2020, DS-GVO Art. 15
  Rn. 86, beck-online
* Paal/Pauly/Paal, 3. Aufl. 2021, DS-GVO Art. 15 Rn. 23
* Paal/Pauly/Paal, 3. Aufl. 2021, DS-GVO Art. 15 Rn. 33